+++ app/controllers/projects_controller.rb
@@ -5,6 +5,9 @@ class ProjectsController < ApplicationController
   def index
     authorize Project
     @pagy, @projects = pagy(policy_scope(Project.open))
+
+    EstablishmentsFetcher.new.call(@projects)
+    IncomingTrustsFetcher.new.call(@projects)
   end
 
   def completed